requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-heroku==0.3.1
django-on-heroku==1.1.2
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
gunicorn==21.2.0
idna==3.4
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
whitenoise==6.6.0
